=doc another batch of documentation (example) fixes
