+++ akka-docs-dev/rst/scala/http/routing-dsl/directives/basic-directives/extract.rst
@@ -23,5 +23,5 @@ See :ref:`ProvideDirectives` for an overview of similar directives.
 Example
 -------
 
-... includecode2:: ../../../../code/docs/http/scaladsl/server/directives/BasicDirectivesExamplesSpec.scala
+.. includecode2:: ../../../../code/docs/http/scaladsl/server/directives/BasicDirectivesExamplesSpec.scala
    :snippet: 0extract
