+++ akka-docs-dev/rst/scala/http/routing-dsl/directives/basic-directives/extractRequest.rst
@@ -3,7 +3,7 @@
 extractRequest
 ==============
 
-...
+Extracts the complete ``HttpRequest`` instance.
 
 Signature
 ---------
@@ -14,10 +14,12 @@ Signature
 Description
 -----------
 
-...
+Use ``extractRequest`` to extract just the complete URI of the request. Usually there's little use of
+extracting the complete request because extracting of most of the aspects of HttpRequests is handled by specialized
+directives. See :ref:`Request Directives`.
 
 Example
 -------
 
-... includecode2:: ../../../../code/docs/http/scaladsl/server/directives/BasicDirectivesExamplesSpec.scala
-   :snippet: 0extractRequest
+.. includecode2:: ../../../../code/docs/http/scaladsl/server/directives/BasicDirectivesExamplesSpec.scala
+   :snippet: extractRequest-example
