+++ akka-docs-dev/rst/scala/http/routing-dsl/directives/basic-directives/extractUri.rst
@@ -3,7 +3,7 @@
 extractUri
 ==========
 
-...
+Access the full URI of the request.
 
 Signature
 ---------
@@ -14,10 +14,11 @@ Signature
 Description
 -----------
 
-...
+Use :ref:`SchemeDirectives`, :ref:`HostDirectives`, :ref:`PathDirectives`,  and :ref:`ParameterDirectives` for more
+targeted access to parts of the URI.
 
 Example
 -------
 
-... includecode2:: ../../../../code/docs/http/scaladsl/server/directives/BasicDirectivesExamplesSpec.scala
-   :snippet: 0extractUri
+.. includecode2:: ../../../../code/docs/http/scaladsl/server/directives/BasicDirectivesExamplesSpec.scala
+   :snippet: extractUri-example
