+++ akka-docs-dev/rst/scala/http/routing-dsl/directives/basic-directives/mapRouteResult.rst
@@ -22,5 +22,5 @@ See :ref:`Result Transformation Directives` for similar directives.
 Example
 -------
 
-... includecode2:: ../../../../code/docs/http/scaladsl/server/directives/BasicDirectivesExamplesSpec.scala
+.. includecode2:: ../../../../code/docs/http/scaladsl/server/directives/BasicDirectivesExamplesSpec.scala
    :snippet: 0mapRouteResult
