+++ akka-docs-dev/rst/scala/http/routing-dsl/directives/basic-directives/mapUnmatchedPath.rst
@@ -3,7 +3,7 @@
 mapUnmatchedPath
 ================
 
-...
+Transforms the unmatchedPath field of the request context for inner routes.
 
 Signature
 ---------
@@ -14,10 +14,13 @@ Signature
 Description
 -----------
 
-...
+The ``mapUnmatchedPath`` directive is used as a building block for writing :ref:`Custom Directives`. You can use it
+for implementing custom path matching directives.
+
+Use ``extractUnmatchedPath`` for extracting the current value of the unmatched path.
 
 Example
 -------
 
-... includecode2:: ../../../../code/docs/http/scaladsl/server/directives/BasicDirectivesExamplesSpec.scala
-   :snippet: 0mapUnmatchedPath
+.. includecode2:: ../../../../code/docs/http/scaladsl/server/directives/BasicDirectivesExamplesSpec.scala
+   :snippet: mapUnmatchedPath-example
