+++ akka-docs-dev/rst/scala/http/routing-dsl/directives/form-field-directives/formFields.rst
@@ -92,7 +92,7 @@ to provide a ``Deserializer[Option[BodyPart], T]``.
 Example
 -------
 
-... includecode2:: ../../../../code/docs/http/scaladsl/server/directives/FormFieldDirectivesExamplesSpec.scala
+.. includecode2:: ../../../../code/docs/http/scaladsl/server/directives/FormFieldDirectivesExamplesSpec.scala
    :snippet: formFields
 
 For more examples about the way how fields can specified see the examples for the ``parameters`` directive.
