+++ akka-docs-dev/rst/scala/http/routing-dsl/directives/future-directives/onSuccess.rst
@@ -3,7 +3,8 @@
 onSuccess
 =========
 
-...
+Evaluates its parameter of type ``Future[T]``, and once the ``Future`` has been completed successfully,
+extracts its result as a value of type ``T`` and passes it to the inner route.
 
 Signature
 ---------
@@ -14,10 +15,14 @@ Signature
 Description
 -----------
 
-...
+The execution of the inner route passed to a onSuccess directive is deferred until the given future
+has completed successfully, exposing the future's value as a extraction of type ``T``. If the future
+fails its failure throwable is bubbled up to the nearest ``ExceptionHandler``.
+
+To handle the ``Failure`` case manually as well, use :ref:`-onComplete-`, instead.
 
 Example
 -------
 
-... includecode2:: ../../../../code/docs/http/scaladsl/server/directives/FutureDirectivesExamplesSpec.scala
-   :snippet: 0onSuccess
+.. includecode2:: ../../../../code/docs/http/scaladsl/server/directives/FutureDirectivesExamplesSpec.scala
+   :snippet: onSuccess
