+++ akka-docs-dev/rst/scala/http/routing-dsl/directives/header-directives/headerValue.rst
@@ -26,5 +26,5 @@ syntactic alternative.
 Example
 -------
 
-... includecode2:: ../../../../code/docs/http/scaladsl/server/directives/HeaderDirectivesExamplesSpec.scala
+.. includecode2:: ../../../../code/docs/http/scaladsl/server/directives/HeaderDirectivesExamplesSpec.scala
    :snippet: headerValue-0
