+++ akka-docs-dev/rst/scala/http/routing-dsl/directives/header-directives/headerValueByType.rst
@@ -28,5 +28,5 @@ directive instead.
 Example
 -------
 
-... includecode2:: ../../../../code/docs/http/scaladsl/server/directives/HeaderDirectivesExamplesSpec.scala
+.. includecode2:: ../../../../code/docs/http/scaladsl/server/directives/HeaderDirectivesExamplesSpec.scala
    :snippet: headerValueByType-0
