+++ akka-docs-dev/rst/scala/http/routing-dsl/directives/header-directives/optionalHeaderValueByType.rst
@@ -26,5 +26,5 @@ an ``Option`` value instead of rejecting the request if no matching header could
 Example
 -------
 
-... includecode2:: ../../../../code/docs/http/scaladsl/server/directives/HeaderDirectivesExamplesSpec.scala
+.. includecode2:: ../../../../code/docs/http/scaladsl/server/directives/HeaderDirectivesExamplesSpec.scala
    :snippet: optionalHeaderValueByType-0
