+++ akka-docs-dev/rst/scala/http/routing-dsl/directives/marshalling-directives/completeWith.rst
@@ -32,15 +32,15 @@ The following example uses ``spray-json`` to marshall a simple ``Person`` class 
 response.  It utilizes ``SprayJsonSupport`` via the ``PersonJsonSupport`` object as the in-scope
 unmarshaller.
 
-... includecode2:: ../../../../code/docs/http/scaladsl/server/directives/MarshallingDirectivesExamplesSpec.scala
+.. includecode2:: ../../../../code/docs/http/scaladsl/server/directives/MarshallingDirectivesExamplesSpec.scala
    :snippet: person-json-support 
 
-   ... includecode2:: ../../../../code/docs/http/scaladsl/server/directives/MarshallingDirectivesExamplesSpec.scala
+.. includecode2:: ../../../../code/docs/http/scaladsl/server/directives/MarshallingDirectivesExamplesSpec.scala
    :snippet: person-case-class
 
 The ``findPerson`` takes an argument of type ``Person => Unit`` which is generated by the ``completeWith``
 call.  We can handle any logic we want in ``findPerson`` and call our completion function to
 complete the request.
 
-... includecode2:: ../../../../code/docs/http/scaladsl/server/directives/MarshallingDirectivesExamplesSpec.scala
+.. includecode2:: ../../../../code/docs/http/scaladsl/server/directives/MarshallingDirectivesExamplesSpec.scala
    :snippet: example-completeWith-with-json
