+++ akka-docs-dev/rst/scala/http/routing-dsl/directives/method-directives/patch.rst
@@ -23,5 +23,5 @@ by the default :ref:`RejectionHandler <The RejectionHandler>`.
 Example
 -------
 
-... includecode2:: ../../../../code/docs/http/scaladsl/server/directives/MethodDirectivesExamplesSpec.scala
+.. includecode2:: ../../../../code/docs/http/scaladsl/server/directives/MethodDirectivesExamplesSpec.scala
   :snippet: patch-method
