+++ akka-docs-dev/rst/scala/http/routing-dsl/directives/misc-directives/validate.rst
@@ -16,5 +16,5 @@ Signature
 Example
 -------
 
-... includecode2:: ../../../../code/docs/http/scaladsl/server/directives/MiscDirectivesExamplesSpec.scala
+.. includecode2:: ../../../../code/docs/http/scaladsl/server/directives/MiscDirectivesExamplesSpec.scala
   :snippet: validate-example
